Split Shade utilities into DS and app lanes
ref DES-1325
Replaced per-app lint deny-lists with physical utility module boundaries.
