+++ apps/shade/src/lib/app-utils.ts
@@ -1,0 +1,416 @@
+import moment, {Moment} from 'moment-timezone';
+import isEmail from 'validator/es/lib/isEmail';
+
+import {formatDisplayDate} from './ds-utils';
+
+// Check if string is a domain
+export const isValidDomain = (value: string) => {
+    return /^[a-z0-9](?:[a-z0-9-]*[a-z0-9])?(\.[a-z0-9](?:[a-z0-9-]*[a-z0-9])?)+(?:\/[\w-./?%&=]*)?$/i.test(value);
+};
+
+// Helper to format a URL
+export const formatUrl = (value: string, baseUrl?: string, nullable?: boolean) => {
+    if (nullable && !value) {
+        return {save: null, display: ''};
+    }
+
+    let url = value.trim();
+
+    if (!url) {
+        if (baseUrl) {
+            return {save: '/', display: baseUrl};
+        }
+        return {save: '', display: ''};
+    }
+
+    // if we have an email address, add the mailto:
+    if (isEmail(url)) {
+        return {save: `mailto:${url}`, display: `mailto:${url}`};
+    }
+
+    const isAnchorLink = url.match(/^#/);
+    if (isAnchorLink) {
+        return {save: url, display: url};
+    }
+
+    const isProtocolRelative = url.match(/^(\/\/)/);
+    if (isProtocolRelative) {
+        return {save: url, display: url};
+    }
+
+    if (!baseUrl) {
+        // Absolute URL with no base URL
+        if (!url.startsWith('http')) {
+            url = `https://${url}`;
+        }
+    }
+
+    // If it doesn't look like a URL, leave it as is rather than assuming it's a pathname etc
+    if (!url.match(/^[a-zA-Z0-9-]+:/) && !url.match(/^(\/|\?)/)) {
+        return {save: url, display: url};
+    }
+
+    let parsedUrl: URL;
+
+    try {
+        parsedUrl = new URL(url, baseUrl);
+    } catch {
+        return {save: url, display: url};
+    }
+
+    if (!baseUrl) {
+        return {save: parsedUrl.toString(), display: parsedUrl.toString()};
+    }
+    const parsedBaseUrl = new URL(baseUrl);
+
+    let isRelativeToBasePath = parsedUrl.pathname && parsedUrl.pathname.indexOf(parsedBaseUrl.pathname) === 0;
+
+    // if our path is only missing a trailing / mark it as relative
+    if (`${parsedUrl.pathname}/` === parsedBaseUrl.pathname) {
+        isRelativeToBasePath = true;
+    }
+
+    const isOnSameHost = parsedUrl.host === parsedBaseUrl.host;
+
+    // if relative to baseUrl, remove the base url before sending to action
+    if (isOnSameHost && isRelativeToBasePath) {
+        url = url.replace(/^[a-zA-Z0-9-]+:/, '');
+        url = url.replace(/^\/\//, '');
+        url = url.replace(parsedBaseUrl.host, '');
+        url = url.replace(parsedBaseUrl.pathname, '');
+
+        if (!url.match(/^\//)) {
+            url = `/${url}`;
+        }
+    }
+
+    if (!url.match(/\/$/) && !url.match(/[.#?]/)) {
+        url = `${url}/`;
+    }
+
+    // we update with the relative URL but then transform it back to absolute
+    // for the input value. This avoids problems where the underlying relative
+    // value hasn't changed even though the input value has
+    return {save: url, display: displayFromBase(url, baseUrl)};
+};
+
+// Helper to display a URL from a base URL
+const displayFromBase = (url: string, baseUrl: string) => {
+    // Ensure base url has a trailing slash
+    if (!baseUrl.endsWith('/')) {
+        baseUrl += '/';
+    }
+
+    // Remove leading slash from url
+    if (url.startsWith('/')) {
+        url = url.substring(1);
+    }
+
+    return new URL(url, baseUrl).toString();
+};
+
+// Format date for stats query
+export const formatQueryDate = (date: Moment) => {
+    return date.format('YYYY-MM-DD');
+};
+
+// Format cents to Dollars
+export const centsToDollars = (value: number) => {
+    return Math.round(value / 100);
+};
+
+/* Chart formatters
+/* -------------------------------------------------------------------------- */
+
+// Calculates the Y-axis range with padding
+export const getYRangeWithLargePadding = (data: { value: number }[]): {min: number; max: number} => {
+    if (!data.length) {
+        return {min: 0, max: 1};
+    }
+
+    const values = data.map(d => Number(d.value));
+    let min = Math.min(...values);
+    let max = Math.max(...values);
+
+    // Helper function to round to nearest multiple of 10^n
+    const roundToNearestMultiple = (num: number): number => {
+        if (num === 0) {
+            return 0;
+        }
+
+        // Determine the order of magnitude (10^n)
+        const magnitude = Math.floor(Math.log10(num));
+        const multiple = Math.pow(10, magnitude);
+
+        // Round to nearest multiple
+        return Math.round(num / multiple) * multiple;
+    };
+
+    // Add padding based on magnitude before rounding
+    const magnitude = Math.floor(Math.log10(Math.max(max, 1)));
+    const padding = Math.pow(10, magnitude);
+
+    // Add padding and ensure min is not negative
+    min = Math.max(0, min - padding);
+    max = max + padding;
+
+    // Round to nearest multiple of 10^n
+    min = roundToNearestMultiple(min);
+    max = roundToNearestMultiple(max);
+
+    return {min, max};
+};
+
+export const getYRange = (data: { value: number }[]): {min: number; max: number} => {
+    if (!data.length) {
+        return {min: 0, max: 1};
+    }
+
+    const values = data.map(d => Number(d.value));
+    let min = Math.min(...values);
+    let max = Math.max(...values);
+
+    if (min === max) {
+        const value = min;
+        return {min: Math.max(0, value - 1), max: value + 1};
+    }
+
+    // Use a percentage-based padding (10% of the range)
+    const padding = 0.02;
+
+    // Add padding and ensure min is not negative
+    min = Math.max(0, min - (min * padding));
+    max = max + (max * padding);
+
+    const range = max - min;
+
+    // Determine the order of magnitude for rounding based on the range
+    const rangeMagnitude = Math.floor(Math.log10(range));
+    // Always round to at least 10s, but use larger steps for bigger ranges
+    // const roundTo = Math.max(10, Math.pow(10, rangeMagnitude));
+    const roundTo = Math.pow(10, rangeMagnitude);
+
+    // Round min and max to the appropriate precision
+    const roundedMax = Math.round(max / roundTo) * roundTo;
+    max = roundedMax < max ? Math.ceil(max / roundTo) * roundTo : roundedMax;
+
+    const roundedMin = Math.round(min / roundTo) * roundTo;
+    min = roundedMin > min ? Math.floor(min / roundTo) * roundTo : roundedMin;
+    min = Math.max(0, min);
+
+    // Ensure we have a visible range even after rounding
+    if (min === max) {
+        const midPoint = (min + max) / 2;
+        const smallRange = Math.max(Math.abs(midPoint) * padding, roundTo);
+        min = Math.max(0, Math.floor(midPoint - smallRange));
+        max = Math.ceil(midPoint + smallRange);
+    }
+
+    // Final safety check to ensure min is never negative
+    min = Math.max(0, min);
+
+    return {min, max};
+};
+
+// Unfortunately in order to force Recharts area charts to start at a certain value
+// we need to use allowDataOverflow = true on the yAxis. This however clips the min
+// value if it reaches 0. In order to prevent this happening we add a bit of padding
+// to the min value.
+export const getYRangeWithMinPadding = (range: {min: number; max: number}) => {
+    if (range.min !== 0) {
+        return [range.min, range.max];
+    }
+    const padding = 0.005;
+    const minPadding = -2;
+    return [Math.min(range.min - (range.max * padding), minPadding), range.max];
+};
+
+// Calculates the width needed for the Y-axis based on the formatted tick values
+export const calculateYAxisWidth = (ticks: number[], formatter: (value: number) => string): number => {
+    if (!ticks.length) {
+        return 40;
+    }
+
+    // Get the longest formatted tick value
+    const maxFormattedLength = Math.max(...ticks.map(tick => formatter(tick).length));
+
+    // Approximate width based on character count (assuming monospace font)
+    // Add padding for safety
+    const width = Math.max(20, maxFormattedLength * 8 + 20);
+    return width;
+};
+
+// Get range for date
+export const getRangeForStartDate = (startDate: string) => {
+    const publishedDate = new Date(startDate);
+    const today = new Date();
+    const diffInTime = today.getTime() - publishedDate.getTime();
+    const diffInDays = Math.ceil(diffInTime / (1000 * 3600 * 24));
+
+    // Ensure minimum of 1 day to avoid issues with same-day publications
+    return Math.max(diffInDays, 1);
+};
+
+//Return today and startdate for charts
+export const getRangeDates = (range: number) => {
+    const timezone = Intl.DateTimeFormat().resolvedOptions().timeZone;
+    const endDate = moment().tz(timezone).endOf('day');
+    let startDate;
+
+    if (range === -1) {
+        // Year to date - use January 1st of current year
+        startDate = moment().tz(timezone).startOf('year');
+    } else {
+        // Regular range calculation
+        startDate = moment().tz(timezone).subtract(range - 1, 'days').startOf('day');
+    }
+
+    return {startDate, endDate, timezone};
+};
+
+/**
+ * Sanitizes chart data based on the date range
+ * - For ranges between 91-356 days: shows weekly changes
+ * - For ranges above 356 days: shows monthly changes
+ * - For other ranges: keeps data as is
+ * @param data The chart data to sanitize
+ * @param range The date range in days
+ * @param fieldName The name of the field to use for calculations
+ * @param aggregationType The type of aggregation to use: 'sum', 'avg', or 'exact'
+ */
+export const sanitizeChartData = <T extends {date: string}>(data: T[], range: number, fieldName: keyof T = 'value' as keyof T, aggregationType: 'sum' | 'avg' | 'exact' = 'avg'): T[] => {
+    if (!data.length) {
+        return [];
+    }
+
+    if (range >= 91 && range <= 356) {
+        // Weekly changes
+        const weeklyData: T[] = [];
+        let currentWeek = moment(data[0].date).startOf('week');
+        let weekTotal = 0;
+        let weekCount = 0;
+        let lastValue = 0;
+
+        data.forEach((item, index) => {
+            const itemDate = moment(item.date);
+            if (itemDate.isSame(currentWeek, 'week')) {
+                weekTotal += Number(item[fieldName]);
+                weekCount += 1;
+                lastValue = Number(item[fieldName]);
+            } else {
+                // Add the value for the previous week
+                weeklyData.push({
+                    ...data[index - 1],
+                    date: currentWeek.format('YYYY-MM-DD'),
+                    [fieldName]: aggregationType === 'sum' ? weekTotal :
+                        aggregationType === 'avg' ? (weekCount > 0 ? weekTotal / weekCount : 0) :
+                            lastValue
+                } as T);
+
+                // Start new week
+                currentWeek = itemDate.startOf('week');
+                weekTotal = Number(item[fieldName]);
+                weekCount = 1;
+                lastValue = Number(item[fieldName]);
+            }
+
+            // Handle the last item
+            if (index === data.length - 1) {
+                weeklyData.push({
+                    ...item,
+                    date: currentWeek.format('YYYY-MM-DD'),
+                    [fieldName]: aggregationType === 'sum' ? weekTotal :
+                        aggregationType === 'avg' ? (weekCount > 0 ? weekTotal / weekCount : 0) :
+                            lastValue
+                } as T);
+            }
+        });
+
+        return weeklyData;
+    } else if (range > 356) {
+        // Monthly changes
+        const monthlyData: T[] = [];
+        let currentMonth = moment(data[0].date).startOf('month');
+        let monthTotal = 0;
+        let monthCount = 0;
+        let lastValue = 0;
+
+        data.forEach((item, index) => {
+            const itemDate = moment(item.date);
+            if (itemDate.isSame(currentMonth, 'month')) {
+                monthTotal += Number(item[fieldName]);
+                monthCount += 1;
+                lastValue = Number(item[fieldName]);
+            } else {
+                // Add the value for the previous month
+                monthlyData.push({
+                    ...data[index - 1],
+                    date: currentMonth.format('YYYY-MM-DD'),
+                    [fieldName]: aggregationType === 'sum' ? monthTotal :
+                        aggregationType === 'avg' ? (monthCount > 0 ? monthTotal / monthCount : 0) :
+                            lastValue
+                } as T);
+
+                // Start new month
+                currentMonth = itemDate.startOf('month');
+                monthTotal = Number(item[fieldName]);
+                monthCount = 1;
+                lastValue = Number(item[fieldName]);
+            }
+
+            // Handle the last item
+            if (index === data.length - 1) {
+                monthlyData.push({
+                    ...item,
+                    date: currentMonth.format('YYYY-MM-DD'),
+                    [fieldName]: aggregationType === 'sum' ? monthTotal :
+                        aggregationType === 'avg' ? (monthCount > 0 ? monthTotal / monthCount : 0) :
+                            lastValue
+                } as T);
+            }
+        });
+
+        return monthlyData;
+    }
+
+    // Return original data for ranges < 91 days
+    return data;
+};
+
+/**
+ * Formats a date based on the range
+ * - For ranges above 365 days: shows month and year (e.g. "Apr 2025")
+ * - For ranges above 91 days: shows "Week of [date]"
+ * - For other ranges: uses the default formatDisplayDate
+ */
+export const formatDisplayDateWithRange = (date: string, range: number, showHours: boolean = false, hoursOnly: boolean = false): string => {
+    if (range === 1 && hoursOnly) {
+        return moment(date).format('h:mma');
+    } else if (range === 1 && showHours) {
+        return moment(date).format('MMM D, h:mma');
+    } else if (range > 365) {
+        return moment(date).format('MMM YYYY');
+    } else if (range >= 91) {
+        return `Week of ${formatDisplayDate(date)}`;
+    }
+    return formatDisplayDate(date);
+};
+
+/**
+ * Member formatters
+ */
+
+// Helper function to format member names with fallback to email
+export const formatMemberName = (member: {name?: string; email?: string}) => {
+    return (member.name && member.name.trim()) || member.email || 'Unknown Member';
+};
+
+// Helper function to get member initials
+export const getMemberInitials = (member: {name?: string}) => {
+    const name = formatMemberName(member);
+    const words = name.split(' ');
+    if (words.length >= 2) {
+        return (words[0][0] + words[words.length - 1][0]).toUpperCase();
+    }
+    return name.substring(0, 2).toUpperCase();
+};
