+++ apps/shade/src/lib/ds-utils.ts
@@ -1,0 +1,250 @@
+import {clsx, type ClassValue} from 'clsx';
+import moment from 'moment-timezone';
+import {twMerge} from 'tailwind-merge';
+
+/* Generic helper functions
+/* -------------------------------------------------------------------------- */
+
+// Helper to merge Tailwind classes
+export function cn(...inputs: ClassValue[]) {
+    return twMerge(clsx(inputs));
+}
+
+// Helper to debounce a function
+export function debounce<T extends unknown[]>(func: (...args: T) => void, wait: number, immediate: boolean = false): (...args: T) => void {
+    let timeoutId: ReturnType<typeof setTimeout> | null;
+
+    return function (this: unknown, ...args: T): void {
+        const later = () => {
+            timeoutId = null;
+            if (!immediate) {
+                func.apply(this, args);
+            }
+        };
+
+        const callNow = immediate && !timeoutId;
+
+        if (timeoutId) {
+            clearTimeout(timeoutId);
+        }
+
+        timeoutId = setTimeout(later, wait);
+
+        if (callNow) {
+            func.apply(this, args);
+        }
+    };
+}
+
+/* Data formatters
+/* -------------------------------------------------------------------------- */
+
+// Helper to convert kebab-case to PascalCase with numbers
+export const kebabToPascalCase = (str: string): string => {
+    const processed = str
+        .replace(/[-_]([a-z0-9])/gi, (_, char) => char.toUpperCase());
+    return processed.charAt(0).toUpperCase() + processed.slice(1);
+};
+
+// Format date for UI, result is in the format of `12 Jun 2025`
+// When timezone is provided, the date will be converted to that timezone before formatting
+export const formatDisplayDate = (dateString: string, timezone?: string): string => {
+    // If the date is a Date object, convert it to a string
+    // @ts-expect-error This should error if dateString is not a string, but for some reason Typescript isn't catching this
+    if (dateString instanceof Date) {
+        dateString = dateString.toISOString();
+    }
+    // Fallback to empty string if dateString is an unexpected type. Better to fallback to empty string than to crash the app
+    if (!dateString || dateString.length === 0 || typeof dateString !== 'string') {
+        return '';
+    }
+
+    // Check if this is a datetime string (contains time) or just a date
+    const hasTime = dateString.includes(':');
+    const isISOFormat = dateString.includes('T') || dateString.includes('Z');
+
+    let day, month, year, isToday, isCurrentYear;
+
+    // If timezone is provided and this is an ISO format date, use moment-timezone to convert
+    if (timezone && isISOFormat) {
+        const dateMoment = moment.tz(dateString, timezone);
+        const todayMoment = moment.tz(timezone);
+
+        day = dateMoment.date();
+        month = dateMoment.month();
+        year = dateMoment.year();
+        isToday = dateMoment.isSame(todayMoment, 'day');
+        isCurrentYear = year === todayMoment.year();
+    } else {
+        const date = new Date(dateString);
+        const today = new Date();
+
+        if (hasTime && !isISOFormat) {
+            // This is a localized datetime string like "2025-07-29 19:00:00"
+            // Use local date methods to avoid timezone conversion
+            day = date.getDate();
+            month = date.getMonth();
+            year = date.getFullYear();
+            isToday = date.toDateString() === today.toDateString();
+            isCurrentYear = year === today.getFullYear();
+        } else {
+            // This is either a date-only string or an ISO format with timezone (without explicit timezone param)
+            // Use UTC methods as before
+            day = date.getUTCDate();
+            month = date.getUTCMonth();
+            year = date.getUTCFullYear();
+            isToday = date.toISOString().slice(0, 10) === today.toISOString().slice(0, 10);
+            isCurrentYear = year === today.getUTCFullYear();
+        }
+    }
+
+    const months = ['Jan', 'Feb', 'Mar', 'Apr', 'May', 'Jun', 'Jul', 'Aug', 'Sep', 'Oct', 'Nov', 'Dec'];
+    const monthName = months[month];
+
+    if (isToday) {
+        return `${day} ${monthName}`;
+    }
+
+    return isCurrentYear ? `${day} ${monthName}` : `${day} ${monthName} ${year}`;
+};
+
+/**
+ * Format a plain time in a given time zone.
+ *
+ * @example
+ * formatDisplayTime('2020-04-20T18:09:12.345Z', 'Africa/Lagos')
+ * // 7:09pm
+ */
+export const formatDisplayTime = (dateString: string, timezone: string): string => (
+    moment(dateString).tz(timezone).format('h:mma')
+);
+
+// Helper function to format timestamp
+export const formatTimestamp = (timestamp: string) => {
+    const date = new Date(timestamp);
+    const now = new Date();
+
+    // Handle invalid dates
+    if (isNaN(date.getTime())) {
+        return 'Unknown';
+    }
+
+    // Both dates are now in the same timezone context (local)
+    // The timestamp from DB is UTC but Date constructor handles the conversion
+    const diffMs = now.getTime() - date.getTime();
+
+    // Handle negative differences (future dates)
+    if (diffMs < 0) {
+        return 'Just now';
+    }
+
+    const diffMins = Math.floor(diffMs / (1000 * 60));
+    const diffHours = Math.floor(diffMs / (1000 * 60 * 60));
+    const diffDays = Math.floor(diffMs / (1000 * 60 * 60 * 24));
+
+    if (diffMins < 1) {
+        return 'Just now';
+    } else if (diffMins < 60) {
+        return `${diffMins} min ago`;
+    } else if (diffHours < 24) {
+        return `${diffHours} hr ago`;
+    } else if (diffDays === 1) {
+        return 'Yesterday';
+    } else if (diffDays < 7) {
+        return `${diffDays} days ago`;
+    } else {
+        return date.toLocaleDateString('en-US', {
+            month: 'short',
+            day: 'numeric',
+            year: diffDays > 365 ? 'numeric' : undefined
+        });
+    }
+};
+
+// Add thousands indicator to numbers
+export const formatNumber = (value: number): string => {
+    if (isNaN(value) || !isFinite(value)) {
+        return '0';
+    }
+    return new Intl.NumberFormat('en-US').format(Math.round(value));
+};
+
+// Abbreviate numbers
+export function abbreviateNumber(number: number) {
+    const num = Number(number);
+
+    if (num < 1000) {
+        return formatNumber(num);
+    }
+
+    if (num < 1000000) {
+        // For thousands: round to nearest 100 for 10.3k, nearest 1000 for 101k
+        const roundTo = num < 100000 ? 100 : 1000;
+
+        const rounded = Math.round(num / roundTo) * roundTo;
+        const abbreviated = rounded / 1000;
+
+        if (abbreviated === 1000) {
+            return '1M';
+        }
+
+        const formatted = abbreviated % 1 === 0 ? abbreviated.toString() : abbreviated.toFixed(1);
+        return `${formatted}k`;
+    }
+
+    // For millions: round to nearest 100,000 for 1.1M
+    const roundTo = 100000;
+    const rounded = Math.round(num / roundTo) * roundTo;
+    const abbreviated = rounded / 1000000;
+    const formatted = abbreviated % 1 === 0 ? abbreviated.toString() : abbreviated.toFixed(1);
+    return `${formatted}M`;
+}
+
+// Format time duration
+export const formatDuration = (seconds: number): string => {
+    const hours = Math.floor(seconds / 3600);
+    const minutes = Math.floor((seconds % 3600) / 60);
+    const remainingSeconds = Math.floor(seconds % 60);
+
+    if (hours <= 0) {
+        if (minutes <= 0) {
+            return `${remainingSeconds}s`;
+        }
+        return `${minutes}m ${remainingSeconds}s`;
+    }
+
+    return `${hours}h ${minutes}m ${remainingSeconds}s`;
+};
+
+// Format a fraction to percentage
+export const formatPercentage = (value: number) => {
+    const percentage = value * 100;
+    if (percentage === 0) {
+        return '0%';
+    } else if (percentage < 0.1) {
+        return `${percentage.toFixed(2)}%`;
+    } else if (percentage < 1) {
+        return `${percentage.toFixed(1)}%`;
+    }
+    const rounded = Math.round(percentage);
+    return `${new Intl.NumberFormat('en-US').format(rounded)}%`;
+};
+
+// Converts a country code to corresponding flag emoji
+export function getCountryFlag(countryCode:string) {
+    if (!countryCode || countryCode === null || countryCode.toUpperCase() === 'NULL' || countryCode === 'ᴺᵁᴸᴸ' || countryCode === 'ᴺᵁ') {
+        return '🏳️';
+    }
+    return countryCode.toUpperCase().replace(/./g, char => String.fromCodePoint(char.charCodeAt(0) + 127397)
+    );
+}
+
+export const stringToHslColor = (str: string, saturation:string, lightness:string) => {
+    let hash = 0;
+    for (let i = 0; i < str.length; i++) {
+        hash = str.charCodeAt(i) + ((hash << 5) - hash);
+    }
+
+    const h = hash % 360;
+    return 'hsl(' + h + ', ' + saturation + '%, ' + lightness + '%)';
+};
